Tratamento de alguns erros. Ver comentários de uma resenha. Ver resenhas de um usuário

+++ BackEnd/src/services/comentarioServices.js
@@ -1,36 +1,28 @@
 require("dotenv").config();
 const knex = require("../database/index");
 
-    async function createComentario(usuario_id,conteudo,resenha_ID ){
+async function createComentario(usuario_id,conteudo,resenha_ID){
     try{
+        if(usuario_id == null || conteudo == null || resenha_ID == null){
+            throw new Error("Preencha todos os campos adequadamente");
+        }
+        const resenhaExiste = await knex("resenha").select("*").where({id: resenha_ID});
+        if(!resenhaExiste){
+            throw new Error("Resenha não existe!");
+        }
         const comentario= {
             usuario_ID: usuario_id, 
             conteudo: conteudo,
             resenha_ID: resenha_ID
         }
-            await knex("comentario").insert(comentario);
-
-            return "Comentário enviado."
-    
+        await knex("comentario").insert(comentario);
+        return "Comentário enviado."
     }catch(erro){
         throw erro;
-
     }
 }
-    async function lerComentario(){
-    try{
-        const comentarios = await knex("comentario").select("*");
-        if (comentarios.length === 0){
-            throw new Error("Sem comentarios no registro.");
-        }
-        
-        return comentarios;
 
-    }catch(erro){
-        throw(erro);
-    }
-}
-    async function lerComentarioPorId(id){
+async function lerComentarioPorId(id){
     try{
         const comentario= await knex("comentario").select("*").where({id: id}).first();
 
@@ -42,35 +34,51 @@ const knex = require("../database/index");
         throw erro;
     }
 }
-    async function atualizarComentario(id, conteudo){
+
+async function comentariosDeUmaResenha(resenha_id){
+    try{
+        const comentario = await knex("comentario").select("*").where({resenha_ID: resenha_id});
+        if(comentario.length === 0){
+            throw new Error("Essa resenha não possui comentários");
+        }
+        return comentario;
+    }catch(erro){
+        throw erro;
+    }
+}
+
+async function atualizarComentario(id_usuario, id, conteudo){
     try{
         const comentario = await knex("comentario").select("*").where({id:id}).first();
         if(!comentario){
             throw new Error("Comentário não existe");
         }
+        if(comentario.usuario_ID !== id_usuario){
+            throw new Error("Esse comentário não pertence a esse usuário");
+        }
+
         const novoComentario = {
-        conteudo: conteudo,
-        
+            conteudo: conteudo,
         }
 
         await knex("comentario").update(novoComentario).where({id:id});
-
         return "Comentário atualizado."
-
     }catch(erro){
         throw erro;
     }
 }
-    async function deletarComentario(id){
+
+async function deletarComentario(id_usuario, id){
     try{
         const comentario = await knex("comentario").select("*").where({id:id}).first();
-
-            if(!comentario){
-                throw new Error("Comentario não existe.");
-            }
-    
-            await knex("comentario").delete().where({id:id});    
-            return "Comentario deletado."
+        if(!comentario){
+            throw new Error("Comentario não existe.");
+        }
+        if(comentario.usuario_ID !== id_usuario){
+            throw new Error("O comentário não pertence a esse usuário");
+        }
+        await knex("comentario").delete().where({id:id});    
+        return "Comentario deletado."
     
     }catch(erro){
         throw erro;
@@ -79,8 +87,8 @@ const knex = require("../database/index");
 
 module.exports = {
     createComentario,
-    lerComentario,
     lerComentarioPorId,
     atualizarComentario,
-    deletarComentario
+    deletarComentario,
+    comentariosDeUmaResenha
 }
